+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+
+### Fixed
+- Fullpage screenshot with RecyclerView on Android. [Trello 1416](https://trello.com/c/cpHFs1Zr), [Trello 953](https://trello.com/c/3UOEHgCl)
+
 ## [4.2.1] - 2019-11-1
 ### Fixed
 - Send viewport screenshot on each locators call. [Trello 1208](https://trello.com/c/tDVd9f7K/1208-visual-locators-returns-empty-locator)